Documentation and additional refinements
