+++ Sources/TranslationCatalogIO/RenderFormat.swift
@@ -1,3 +1,4 @@
+/// File types which can be generated containing `Expression`/`Translation` info.
 public enum RenderFormat: String, CaseIterable {
     case html
     case markdown
